+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - `linalg:` Ported `BLAS` `spr()`
 - `linalg:` Ported `Matrix` `copy()` and `toString()`, and implemented `clone()` to make matrices usable within RDD functions as a zero value parameter
 - `linalg:` Ported `RowMatrix` `computeGramianMatrix()` and `triuToFull()`
+- `util:` New utility equivalent to `scala.collection.Iterator` `fill()`
 
 ### Changed
 - `linalg:` Use native Lua 1-based `colPtrs` indexes in `SparseMatrix`